# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CreateDecisionRequestPriority =
  (typeof CreateDecisionRequestPriority)[keyof typeof CreateDecisionRequestPriority];

export const CreateDecisionRequestPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export type CreateDecisionRequestStressLevel =
  (typeof CreateDecisionRequestStressLevel)[keyof typeof CreateDecisionRequestStressLevel];

export const CreateDecisionRequestStressLevel = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export type CreateDecisionRequestDeadline =
  (typeof CreateDecisionRequestDeadline)[keyof typeof CreateDecisionRequestDeadline];

export const CreateDecisionRequestDeadline = {
  none: "none",
  today: "today",
  this_week: "this_week",
  later: "later",
} as const;

export type CreateDecisionRequestEnergyLevel =
  (typeof CreateDecisionRequestEnergyLevel)[keyof typeof CreateDecisionRequestEnergyLevel];

export const CreateDecisionRequestEnergyLevel = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export type CreateDecisionRequestImportance =
  (typeof CreateDecisionRequestImportance)[keyof typeof CreateDecisionRequestImportance];

export const CreateDecisionRequestImportance = {
  minor: "minor",
  moderate: "moderate",
  major: "major",
  life_changing: "life_changing",
} as const;

export type CreateDecisionRequestBudgetImpact =
  (typeof CreateDecisionRequestBudgetImpact)[keyof typeof CreateDecisionRequestBudgetImpact];

export const CreateDecisionRequestBudgetImpact = {
  none: "none",
  small: "small",
  medium: "medium",
  large: "large",
} as const;

export type CreateDecisionRequestSocialInfluence =
  (typeof CreateDecisionRequestSocialInfluence)[keyof typeof CreateDecisionRequestSocialInfluence];

export const CreateDecisionRequestSocialInfluence = {
  just_me: "just_me",
  close_family: "close_family",
  team: "team",
  public: "public",
} as const;

export type CreateDecisionRequestGutFeeling =
  (typeof CreateDecisionRequestGutFeeling)[keyof typeof CreateDecisionRequestGutFeeling];

export const CreateDecisionRequestGutFeeling = {
  go_for_it: "go_for_it",
  unsure: "unsure",
  avoid_it: "avoid_it",
} as const;

export type CreateDecisionRequestCategory =
  (typeof CreateDecisionRequestCategory)[keyof typeof CreateDecisionRequestCategory];

export const CreateDecisionRequestCategory = {
  work: "work",
  health: "health",
  finance: "finance",
  relationships: "relationships",
  personal: "personal",
  other: "other",
} as const;

export interface CreateDecisionRequest {
  /** @minLength 3 */
  problem: string;
  mood?: string;
  timeAvailable?: string;
  priority?: CreateDecisionRequestPriority;
  /**
   * @minimum 0
   * @maximum 24
   */
  sleepHours?: number;
  stressLevel?: CreateDecisionRequestStressLevel;
  deadline?: CreateDecisionRequestDeadline;
  energyLevel?: CreateDecisionRequestEnergyLevel;
  importance?: CreateDecisionRequestImportance;
  budgetImpact?: CreateDecisionRequestBudgetImpact;
  socialInfluence?: CreateDecisionRequestSocialInfluence;
  gutFeeling?: CreateDecisionRequestGutFeeling;
  category?: CreateDecisionRequestCategory;
  useAi?: boolean;
}

export type UpdateOutcomeRequestOutcome =
  (typeof UpdateOutcomeRequestOutcome)[keyof typeof UpdateOutcomeRequestOutcome];

export const UpdateOutcomeRequestOutcome = {
  great: "great",
  okay: "okay",
  regret: "regret",
} as const;

export interface UpdateOutcomeRequest {
  outcome: UpdateOutcomeRequestOutcome;
  outcomeNote?: string;
}

export interface Decision {
  id: number;
  problem: string;
  mood?: string | null;
  timeAvailable?: string | null;
  priority?: string | null;
  sleepHours?: number | null;
  stressLevel?: string | null;
  deadline?: string | null;
  energyLevel?: string | null;
  importance?: string | null;
  budgetImpact?: string | null;
  socialInfluence?: string | null;
  gutFeeling?: string | null;
  category?: string | null;
  outcome?: string | null;
  outcomeNote?: string | null;
  ruleDecision: string;
  ruleExplanation: string;
  finalDecision: string;
  explanation: string;
  confidence: number;
  aiUsed: boolean;
  createdAt: string;
}

export type DecisionDashboardOutcomeStats = {
  great: number;
  okay: number;
  regret: number;
  pending: number;
};

export type DecisionDashboardCategoryBreakdownItem = {
  label: string;
  count: number;
};

export type DecisionDashboardMoodBreakdownItem = {
  label: string;
  count: number;
};

export interface DecisionDashboard {
  todayCount: number;
  totalCount: number;
  aiEnhancedCount: number;
  averageConfidence: number;
  streak: number;
  outcomeStats: DecisionDashboardOutcomeStats;
  patternInsights: string[];
  categoryBreakdown: DecisionDashboardCategoryBreakdownItem[];
  recentDecisions: Decision[];
  moodBreakdown: DecisionDashboardMoodBreakdownItem[];
}

export type ListDecisionsParams = {
  search?: string;
  mood?: string;
  priority?: ListDecisionsPriority;
  category?: string;
};

export type ListDecisionsPriority =
  (typeof ListDecisionsPriority)[keyof typeof ListDecisionsPriority];

export const ListDecisionsPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;