# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: decisions
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /decisions:
    get:
      operationId: listDecisions
      tags: [decisions]
      summary: List saved decisions
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: mood
          in: query
          required: false
          schema:
            type: string
        - name: priority
          in: query
          required: false
          schema:
            type: string
            enum: [low, medium, high]
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Saved decisions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Decision"
    post:
      operationId: createDecision
      tags: [decisions]
      summary: Create a decision recommendation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDecisionRequest"
      responses:
        "201":
          description: Created decision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Decision"
  /decisions/dashboard:
    get:
      operationId: getDecisionDashboard
      tags: [decisions]
      summary: Decision dashboard summary
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecisionDashboard"
  /decisions/{id}:
    delete:
      operationId: deleteDecision
      tags: [decisions]
      summary: Delete a saved decision
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /decisions/{id}/outcome:
    patch:
      operationId: updateDecisionOutcome
      tags: [decisions]
      summary: Update outcome of a past decision
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOutcomeRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Decision"
components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
    CreateDecisionRequest:
      type: object
      required: [problem]
      properties:
        problem:
          type: string
          minLength: 3
        mood:
          type: string
        timeAvailable:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
        sleepHours:
          type: number
          minimum: 0
          maximum: 24
        stressLevel:
          type: string
          enum: [low, medium, high]
        deadline:
          type: string
          enum: [none, today, this_week, later]
        energyLevel:
          type: string
          enum: [low, medium, high]
        importance:
          type: string
          enum: [minor, moderate, major, life_changing]
        budgetImpact:
          type: string
          enum: [none, small, medium, large]
        socialInfluence:
          type: string
          enum: [just_me, close_family, team, public]
        gutFeeling:
          type: string
          enum: [go_for_it, unsure, avoid_it]
        category:
          type: string
          enum: [work, health, finance, relationships, personal, other]
        useAi:
          type: boolean
    UpdateOutcomeRequest:
      type: object
      required: [outcome]
      properties:
        outcome:
          type: string
          enum: [great, okay, regret]
        outcomeNote:
          type: string
    Decision:
      type: object
      required: [id, problem, ruleDecision, ruleExplanation, finalDecision, explanation, confidence, aiUsed, createdAt]
      properties:
        id:
          type: integer
        problem:
          type: string
        mood:
          type: string
          nullable: true
        timeAvailable:
          type: string
          nullable: true
        priority:
          type: string
          nullable: true
        sleepHours:
          type: number
          nullable: true
        stressLevel:
          type: string
          nullable: true
        deadline:
          type: string
          nullable: true
        energyLevel:
          type: string
          nullable: true
        importance:
          type: string
          nullable: true
        budgetImpact:
          type: string
          nullable: true
        socialInfluence:
          type: string
          nullable: true
        gutFeeling:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        outcome:
          type: string
          nullable: true
        outcomeNote:
          type: string
          nullable: true
        ruleDecision:
          type: string
        ruleExplanation:
          type: string
        finalDecision:
          type: string
        explanation:
          type: string
        confidence:
          type: integer
        aiUsed:
          type: boolean
        createdAt:
          type: string
          format: date-time
    DecisionDashboard:
      type: object
      required: [todayCount, totalCount, aiEnhancedCount, averageConfidence, recentDecisions, moodBreakdown, streak, outcomeStats, patternInsights, categoryBreakdown]
      properties:
        todayCount:
          type: integer
        totalCount:
          type: integer
        aiEnhancedCount:
          type: integer
        averageConfidence:
          type: integer
        streak:
          type: integer
        outcomeStats:
          type: object
          required: [great, okay, regret, pending]
          properties:
            great:
              type: integer
            okay:
              type: integer
            regret:
              type: integer
            pending:
              type: integer
        patternInsights:
          type: array
          items:
            type: string
        categoryBreakdown:
          type: array
          items:
            type: object
            required: [label, count]
            properties:
              label:
                type: string
              count:
                type: integer
        recentDecisions:
          type: array
          items:
            $ref: "#/components/schemas/Decision"
        moodBreakdown:
          type: array
          items:
            type: object
            required: [label, count]
            properties:
              label:
                type: string
              count:
                type: integer