# lib/api-zod/src/generated/types/createDecisionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateDecisionRequestBudgetImpact } from "./createDecisionRequestBudgetImpact";
import type { CreateDecisionRequestCategory } from "./createDecisionRequestCategory";
import type { CreateDecisionRequestDeadline } from "./createDecisionRequestDeadline";
import type { CreateDecisionRequestEnergyLevel } from "./createDecisionRequestEnergyLevel";
import type { CreateDecisionRequestGutFeeling } from "./createDecisionRequestGutFeeling";
import type { CreateDecisionRequestImportance } from "./createDecisionRequestImportance";
import type { CreateDecisionRequestPriority } from "./createDecisionRequestPriority";
import type { CreateDecisionRequestSocialInfluence } from "./createDecisionRequestSocialInfluence";
import type { CreateDecisionRequestStressLevel } from "./createDecisionRequestStressLevel";

export interface CreateDecisionRequest {
  /** @minLength 3 */
  problem: string;
  mood?: string;
  timeAvailable?: string;
  priority?: CreateDecisionRequestPriority;
  /**
   * @minimum 0
   * @maximum 24
   */
  sleepHours?: number;
  stressLevel?: CreateDecisionRequestStressLevel;
  deadline?: CreateDecisionRequestDeadline;
  energyLevel?: CreateDecisionRequestEnergyLevel;
  importance?: CreateDecisionRequestImportance;
  budgetImpact?: CreateDecisionRequestBudgetImpact;
  socialInfluence?: CreateDecisionRequestSocialInfluence;
  gutFeeling?: CreateDecisionRequestGutFeeling;
  category?: CreateDecisionRequestCategory;
  useAi?: boolean;
}