# lib/api-zod/src/generated/types/decision.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Decision {
  id: number;
  problem: string;
  mood?: string | null;
  timeAvailable?: string | null;
  priority?: string | null;
  sleepHours?: number | null;
  stressLevel?: string | null;
  deadline?: string | null;
  energyLevel?: string | null;
  importance?: string | null;
  budgetImpact?: string | null;
  socialInfluence?: string | null;
  gutFeeling?: string | null;
  category?: string | null;
  outcome?: string | null;
  outcomeNote?: string | null;
  ruleDecision: string;
  ruleExplanation: string;
  finalDecision: string;
  explanation: string;
  confidence: number;
  aiUsed: boolean;
  createdAt: Date;
}